refactor: move other generic helpers to `mpc-utils`
