+++ chain-signatures/chain-ethereum/src/event_parsing.rs
@@ -13,7 +13,7 @@ use mpc_primitives::{
     Chain, ChainEvent, IndexedSignRequest, SignArgs, SignId, Signature as MpcSignature,
     SignatureRespondedEvent, LATEST_MPC_KEY_VERSION, MAX_SECP256K1_SCALAR,
 };
-use mpc_utils::current_unix_timestamp;
+use mpc_utils::time::current_unix_timestamp;
 use tokio::sync::mpsc;
 
 /// Whether a transaction's calldata represents a contract call.
