+++ chain-signatures/node/src/protocol/triple.rs
@@ -7,10 +7,9 @@ use crate::mesh::MeshState;
 use crate::protocol::posit::{PositRejectReason, Positor};
 use crate::storage::triple_storage::{TriplePair, TriplePairSlot, TripleStorage};
 use crate::types::TripleProtocol;
-use crate::util::JoinMap;
 use mpc_chain_near::AffinePointExt as _;
-
 use mpc_contract::config::ProtocolConfig;
+use mpc_utils::task::JoinMap;
 
 use cait_sith::protocol::{Action, InitializationError, Participant};
 use cait_sith::triples::{TriplePub, TripleShare};
