+++ chain-signatures/node/src/respond_bidirectional.rs
@@ -6,7 +6,7 @@ use mpc_primitives::{
     BidirectionalTx, Chain, ChainConfig as _, IndexedSignRequest,
     RespondBidirectionalSerializedOutput, RespondBidirectionalTx, SerDeserFormat, SignArgs, SignId,
 };
-use mpc_utils::current_unix_timestamp;
+use mpc_utils::time::current_unix_timestamp;
 
 const MAGIC_ERROR_PREFIX: [u8; 4] = [0xde, 0xad, 0xbe, 0xef];
 const SOLANA_RESPOND_BIDIRECTIONAL_PATH: &str = "solana response key";
